add  domain dedicated to crypto key operations
