+++ internal/app/services/crypto.go
@@ -1,0 +1,5 @@
+package services
+
+// CryptoKeyOperationService implements the ICryptoKeyOperationService interface for local cryptographic key management, encryption, signing, and PKCS#11 operations.
+type CryptoKeyOperationService struct {
+}
